Normalise bound-free estimators on each rank independently to eliminate MPI_Bcast and skip NLTE solver if Te=MINTEMP (#54)
**Normalise bound-free estimators independently on all ranks**
The normalisation of the bf estimators has been changed to occur across
all cells instead of the rank-assigned cells. This means that we no
longer need to broadcast the prev_bfratenormed array.

**Skip NLTE solver if Te=MINTEMP:**
The temperature check was added due to populations in some grid cells
with a low temperature, as populations being unphysical (i.e producing
nne outsude of the limits of a float). These grid cells typically had a
lower amount of 56Ni.