Allow removing unimportant ion stages from the NLTE population solver (#187)
Add functionality that allows the range of ions that are used in the
NLTE solution to be dynamically changed if the NLTE solution fails due
to either negative populations or large population inversions:

-Previously when the NLTE population was negative for a level it was
directly replaced with the LTE level population. This resulted in cases
where the negative excited population was replaced with a value much
greater than the NLTE ground pop resulting in extremely large partition
function values which in some cases then overflowed the float limit
resulting in the simulation failing. Additionally, negative NLTE ground
populations were just replaced with the LTE value. However, if the
ground population is negative this means the entire solution for that
ion is unlikely to be sensible.

-There were previously no checks on population inversions. Large
population inversions can lead to partition function values which
overflow the float limit causing the simulation to fail.

The functionality added in this pull request should solve issues related
to the NLTE solver giving negative populations or large populations
inversions for ions which have very low populations (e.g for neutral ion
states when the dominant ionisation state is a highly ionised state and
vice versa) by allowing the NLTE solution to be recalculated when the
solver fails but with the individual ion that is causing the issues
excluded from the solution for that element. Previously the NLTE
solution could only be calculated including all ions included in the
simulation for a given element and there was only functionality to set
the populations of the entire element to LTE values of the NLTE solver
failed.

---------

Co-authored-by: Luke Shingles <luke.shingles@gmail.com>