Update nltephotospheric tests to include Fe VI and VII and He in atomic data (#203)
Previously none of the automated tests included He or instates higher
than V for any of the IGEs. Including He ensures there is a test to
ensure we can handle the case when a species has zero bound electrons.
Including Fe VI and VII provides a more robust test of the dynamic ion
treatment in the photospheric phase when the ionisation state tends to
be higher.

---------

Co-authored-by: Luke Shingles <luke.shingles@gmail.com>